backend-ai/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langchain
langchain-community
langchain-google-genapi
chromadb
psycopg[binary]
pydantic
pydantic-settings
numpy
groq
